% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qc_tables.R
\name{make_nice_table}
\alias{make_nice_table}
\title{Make a nice data.frame}
\usage{
make_nice_table(data_df = NULL, sequence = c("all", "per"))
}
\arguments{
\item{data_df}{the data frame}

\item{sequence}{does the table show hte qc info for all sequences or per sequence}
}
\value{
nice data.frame
}
\description{
Make a nice data.frame which can be shown as a nice table.
}
\details{
qc_tables
}
\author{
Rico Derks
}
