% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_qc_graphs.R
\name{make_qc_graph}
\alias{make_qc_graph}
\title{Make the QC graph}
\usage{
make_qc_graph(data_df = NULL, yaxis = NULL, sequence = c("all", "per"))
}
\arguments{
\item{data_df}{data.frame with the data}

\item{yaxis}{what to show on the y-axis}

\item{sequence}{show qc results for all sequences summarised or per sequence}
}
\value{
a gpplot2 object
}
\description{
Make the QC graph
}
\details{
qc_graphs
}
\author{
Rico Derks
}
