% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_in_browser}
\alias{run_in_browser}
\title{Run the app in the broweser}
\usage{
run_in_browser()
}
\value{

}
\description{
Run the app in the browser when started manually from Rstudio
}
\details{
run_in_browser - shortcut to launch in_browser
}
